ai_voice/requirements.txt
paho-mqtt
google-generativeai
python-dotenv